The numbers copied to the clipboard are formated using the système locale
It makes it possible to copy the clipboard content to that dumb Excel that
can't let the user select the number format.

A settings menu was added to disable the number formating and always
output the numbers with a decimal dot.

The copy functions have been moved to a single function instead of
duplicating the code in every function.