Avoid an infinite loop if the y axis is very narrow
If the difference between the min and max is around 1e17 times smaller
than the absolute value, the loop never ends because the step doesn't add
to the iterated value due to the numerical precision of the processor.

I tried to make the displayed values on the axis look better.